rename addition module to add so matches file and function
